// models/items.ts
//  export interface ______ {

// }

export interface Item {
  id: number
  name: string
  description: string
  image: Text
  tradable: boolean
}

export interface ItemData {
  name: string
  description: string
  image: Text
  tradable: boolean
}

// returns [{id: number , name: string, description: string, image: null, tradable: boolean}]